requirements/development.txt
-r production.txt
pep8
ipython[all]
django-debug-toolbar
django-extensions
dj-database-url
django-storages-redux
boto
